+++ config/webpack/shared.js
@@ -7,8 +7,7 @@ const yaml = require('./loaders/yaml');
 const I18n = require('./loaders/I18n');
 const modulejs = require('./loaders/modulejs');
 const tooltip = require('./loaders/tooltip');
-
-const assetsPath = path.resolve(__dirname, '../../public/assets/');
+const SymlinkAssets = require('./plugins/symlink');
 
 environment.plugins.prepend('Provide', new webpack.ProvidePlugin({
   $:      'jquery',
@@ -18,9 +17,7 @@ environment.plugins.prepend('Provide', new webpack.ProvidePlugin({
 
 environment.plugins.prepend('moment', new webpack.ContextReplacementPlugin(/moment[/\\]locale$/, /ru/));
 
-environment.plugins.prepend('clean', new CleanWebpackPlugin({
-  cleanOnceBeforeBuildPatterns: [assetsPath]
-}));
+environment.plugins.prepend('clean', new CleanWebpackPlugin());
 
 environment.loaders.append('yaml', yaml);
 environment.loaders.append('babel', babel);
@@ -40,5 +37,7 @@ config.resolve.alias = {
   shared:                     path.resolve(__dirname, '../../hbw/app/javascript/packs/hbw/components/shared'),
   messenger:                  path.resolve(__dirname, '../../app/javascript/utils/messenger'),
 };
+
+config.plugins.push(new SymlinkAssets(['hbw.js']));
 
 module.exports = config;